requirements.txt
absl-py==0.2.2
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
coremltools
cycler==0.10.0
enum34==1.1.6
gast==0.2.0
grpcio==1.13.0
h5py==2.8.0
idna==2.8
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==3.0.1
mock==2.0.0
natsort
numpy==1.14.5
pandas
pbr==4.1.1
protobuf==3.6.1
pydicom==1.3.0
pyparsing==2.2.2
python-dateutil==2.7.5
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.1.0
SimpleITK
six==1.10.0
sklearn==0.0
soupsieve==1.9.1
tensorboard==1.12.2
tensorflow==1.12.3
termcolor==1.1.0
virtualenv==16.0.0
Werkzeug==0.14.1
